refactor(bot): split playerFactory.ts into focused modules
Extract yt-dlp stream functions to streamBridge.ts and SoundCloud
search/matching logic to soundcloudMatcher.ts. playerFactory.ts retains
only extractor registration and re-exports moved symbols for backward
compatibility. No behavioral changes.
