+++ packages/bot/src/handlers/player/playerFactory.ts
@@ -1,5 +1,5 @@
 import { Player } from 'discord-player'
-import type { BaseExtractor, Track } from 'discord-player'
+import type { BaseExtractor } from 'discord-player'
 import {
     SoundCloudExtractor,
     AppleMusicExtractor,
@@ -8,17 +8,9 @@ import {
 } from '@discord-player/extractor'
 import { SpotifyExtractor } from 'discord-player-spotify'
 import * as playdl from 'play-dl'
-import { spawn } from 'child_process'
-import { PassThrough } from 'stream'
-import type { Readable } from 'stream'
 import type { CustomClient } from '../../types'
-import { errorLog, infoLog, warnLog, debugLog } from '@lucky/shared/utils'
-import {
-    cleanTitle,
-    cleanAuthor,
-    cleanSearchQuery,
-} from '../../utils/music/searchQueryCleaner'
-import { providerHealthService } from '../../utils/music/search/providerHealth'
+import { errorLog, infoLog, warnLog } from '@lucky/shared/utils'
+import { createResilientStream } from './streamBridge'
 
 type CreatePlayerParams = {
     client: CustomClient
@@ -174,366 +166,5 @@ const loadYoutubeExtractor = async (player: Player): Promise<void> => {
     }
 }
 
-/**
- * Stream audio via yt-dlp subprocess.
- *
- * Resolves with the stdout Readable once yt-dlp begins writing data.
- * Rejects with a timeout error if no data arrives within 15 seconds.
- * Exported for testing.
- */
-const ALLOWED_YTDLP_DOMAINS = new Set([
-    'youtube.com',
-    'www.youtube.com',
-    'youtu.be',
-    'music.youtube.com',
-    'soundcloud.com',
-    'www.soundcloud.com',
-])
-
-function validateYtDlpUrl(url: string): void {
-    if (url.startsWith('ytsearch')) return
-    let parsed: URL
-    try {
-        parsed = new URL(url)
-    } catch (error) {
-        throw new Error(`yt-dlp: invalid URL`, { cause: error })
-    }
-    if (parsed.protocol !== 'https:') {
-        throw new Error(`yt-dlp: only https URLs are allowed`)
-    }
-    if (!ALLOWED_YTDLP_DOMAINS.has(parsed.hostname.toLowerCase())) {
-        throw new Error(`yt-dlp: domain not in allowlist: ${parsed.hostname}`)
-    }
-}
-
-export function streamViaYtDlp(url: string): Promise<Readable> {
-    try {
-        validateYtDlpUrl(url)
-    } catch (err) {
-        return Promise.reject(err)
-    }
-    return new Promise<Readable>((resolve, reject) => {
-        const proc = spawn(
-            'yt-dlp',
-            [
-                '--no-playlist',
-                '-f',
-                'bestaudio/best',
-                '-o',
-                '-',
-                '--quiet',
-                '--no-warnings',
-                '--no-progress',
-                // Provide the Node.js runtime so yt-dlp can use the JS extractor
-                // for YouTube. Use process.execPath (the actual running binary)
-                // rather than a hardcoded path so this works on all systems.
-                '--js-runtimes',
-                `node:${process.execPath}`,
-                url,
-            ],
-            { stdio: ['ignore', 'pipe', 'pipe'] },
-        )
-
-        const timeout = setTimeout(() => {
-            proc.kill()
-            reject(new Error('yt-dlp: timed out waiting for stream start'))
-        }, 15_000)
-
-        const stderrChunks: Buffer[] = []
-        proc.stderr!.on('data', (chunk: Buffer) => stderrChunks.push(chunk))
-
-        let settled = false
-
-        proc.stdout!.once('data', (firstChunk: Buffer) => {
-            if (settled) return
-            settled = true
-            clearTimeout(timeout)
-            const through = new PassThrough()
-            through.write(firstChunk)
-            proc.stdout!.pipe(through)
-            resolve(through)
-        })
-
-        proc.once('error', (err) => {
-            if (settled) return
-            settled = true
-            clearTimeout(timeout)
-            reject(err)
-        })
-
-        proc.once('close', (code) => {
-            if (settled) return
-            settled = true
-            clearTimeout(timeout)
-            if (code && code !== 0) {
-                const stderr = Buffer.concat(stderrChunks).toString().trim()
-                const reason = stderr ? ` — ${stderr.split('\n')[0]}` : ''
-                reject(new Error(`yt-dlp exited with code ${code}${reason}`))
-            }
-        })
-    })
-}
-
-/**
- * Search YouTube via yt-dlp using a text query.
- * Delegates to streamViaYtDlp with a ytsearch1: prefix so the spawn logic
- * stays in one place and no new security hotspot is introduced.
- */
-export function streamViaYtDlpSearch(query: string): Promise<Readable> {
-    if (!query.trim())
-        return Promise.reject(new Error('yt-dlp search: empty query'))
-    return streamViaYtDlp(`ytsearch1:${query}`)
-}
-
-/**
- * Bridge fallback chain (discord-player-youtubei v3 createStream signature).
- *
- * Priority: yt-dlp direct URL → yt-dlp YouTube search (Spotify sources) →
- * SoundCloud (full query) → SoundCloud (title only).
- *
- * yt-dlp is tried first for direct URLs (play-dl YouTube streaming is broken
- * due to bot detection). For Spotify-sourced tracks the URL cannot be streamed
- * directly, so a YouTube search via yt-dlp is attempted before falling back to
- * SoundCloud.
- */
-export async function createResilientStream(
-    track: Pick<Track, 'title' | 'author' | 'duration' | 'url'>,
-    _ext?: unknown,
-): Promise<Readable> {
-    const cleanedTitle = cleanTitle(track.title)
-    const cleanedAuthor = cleanAuthor(track.author)
-    const isSpotifyUrl = track.url?.includes('open.spotify.com') ?? false
-
-    debugLog({
-        message: 'Bridge: resolving stream',
-        data: {
-            title: track.title,
-            author: track.author,
-            cleanedTitle,
-            cleanedAuthor,
-            hasUrl: Boolean(track.url),
-            isSpotifyUrl,
-        },
-    })
-
-    if (track.url && !isSpotifyUrl) {
-        try {
-            const stream = await streamViaYtDlp(track.url)
-            infoLog({
-                message: 'Bridge: streamed via yt-dlp',
-                data: { url: track.url, title: cleanedTitle || track.title },
-            })
-            return stream
-        } catch (ytdlpError) {
-            warnLog({
-                message: 'Bridge: yt-dlp failed, falling back to SoundCloud',
-                data: {
-                    error: (ytdlpError as Error).message,
-                    url: track.url,
-                    cleanedTitle,
-                },
-            })
-        }
-    }
-
-    if (isSpotifyUrl) {
-        const ytQuery = `${cleanSearchQuery(cleanedTitle, cleanedAuthor)} official audio`
-        try {
-            const stream = await streamViaYtDlpSearch(ytQuery)
-            infoLog({
-                message:
-                    'Bridge: streamed via yt-dlp YouTube search (Spotify source)',
-                data: { query: ytQuery, title: cleanedTitle },
-            })
-            return stream
-        } catch (ytSearchError) {
-            warnLog({
-                message:
-                    'Bridge: yt-dlp YouTube search failed, falling back to SoundCloud',
-                data: {
-                    error: (ytSearchError as Error).message,
-                    query: ytQuery,
-                    cleanedTitle,
-                },
-            })
-        }
-    }
-
-    if (!providerHealthService.isAvailable('soundcloud')) {
-        warnLog({
-            message:
-                'Bridge: SoundCloud circuit open, skipping fallback stages',
-            data: {
-                title: track.title,
-                cleanedTitle,
-                url: track.url,
-            },
-        })
-        throw new Error(`Bridge exhausted: no stream for "${track.title}"`)
-    }
-
-    try {
-        return await streamViaSoundCloud(
-            cleanSearchQuery(cleanedTitle, cleanedAuthor),
-            track.duration,
-        )
-    } catch (primaryError) {
-        debugLog({
-            message:
-                'Bridge: SoundCloud primary search failed, retrying with title only',
-            data: {
-                error: (primaryError as Error).message,
-                cleanedTitle,
-            },
-        })
-    }
-
-    try {
-        return await streamViaSoundCloud(cleanedTitle, track.duration)
-    } catch (titleOnlyError) {
-        debugLog({
-            message:
-                'Bridge: title-only SoundCloud failed, retrying without parentheticals',
-            data: {
-                error: (titleOnlyError as Error).message,
-                cleanedTitle,
-            },
-        })
-    }
-
-    const openParen = cleanedTitle.indexOf('(')
-    const coreTitle =
-        openParen > 0 ? cleanedTitle.slice(0, openParen).trim() : cleanedTitle
-    if (coreTitle && coreTitle !== cleanedTitle) {
-        try {
-            return await streamViaSoundCloud(coreTitle, track.duration)
-        } catch (coreError) {
-            errorLog({
-                message: 'Bridge: all stages exhausted',
-                error: coreError,
-                data: {
-                    title: track.title,
-                    cleanedTitle,
-                    coreTitle,
-                    url: track.url,
-                    stages: [
-                        'yt-dlp',
-                        'soundcloud-full',
-                        'soundcloud-title',
-                        'soundcloud-core',
-                    ],
-                },
-            })
-        }
-    } else {
-        errorLog({
-            message: 'Bridge: all stages exhausted',
-            data: {
-                title: track.title,
-                cleanedTitle,
-                url: track.url,
-                stages: ['yt-dlp', 'soundcloud-full', 'soundcloud-title'],
-            },
-        })
-    }
-
-    throw new Error(`Bridge exhausted: no stream for "${track.title}"`)
-}
-
-/**
- * Search SoundCloud for the given query and return a streaming handle for the
- * first result whose title and duration validate. Exported for testing.
- */
-export async function streamViaSoundCloud(
-    query: string,
-    trackDuration?: string,
-): Promise<Readable> {
-    if (!query.trim()) {
-        throw new Error('SoundCloud: empty query')
-    }
-
-    const results = await playdl.search(query, {
-        source: { soundcloud: 'tracks' },
-        limit: 5,
-    })
-
-    if (!results.length) {
-        throw new Error(`SoundCloud: no results for "${query}"`)
-    }
-
-    const match = findMatchingSoundCloudResult(query, trackDuration, results)
-    if (!match) {
-        throw new Error(
-            `SoundCloud: no validated match for "${query}" (title/duration mismatch)`,
-        )
-    }
-
-    const scStream = await playdl.stream(match.url)
-    return scStream.stream
-}
-
-type SoundCloudSearchResult = {
-    name: string
-    url: string
-    durationInSec?: number
-}
-
-/**
- * Minimum fraction of query tokens that must appear in the result name.
- * Using 0.75 instead of 1.0 tolerates minor title differences:
- *   - Accent stripping ("Jacaré" → "jacar" vs "Jacare" → "jacare")
- *   - Compound word splits ("DogDog" vs "Dog Dog" → "dog dog")
- *   - DJ credits missing or abbreviated on SoundCloud
- */
-const TITLE_MATCH_THRESHOLD = 0.75
-
-/**
- * Pure match logic, exported so tests can cover it without spinning up play-dl.
- *
- * Requires at least TITLE_MATCH_THRESHOLD fraction of query tokens to appear
- * in the result name, and duration within 30 seconds when available.
- */
-export function findMatchingSoundCloudResult(
-    query: string,
-    trackDuration: string | undefined,
-    results: readonly SoundCloudSearchResult[],
-): SoundCloudSearchResult | undefined {
-    const queryNorm = normalizeForMatch(query)
-    if (!queryNorm) return undefined
-
-    const tokens = queryNorm.split(/ +/).filter(Boolean)
-    if (tokens.length === 0) return undefined
-
-    const trackSec = parseDurationString(trackDuration)
-
-    return results.find((result) => {
-        const resultNorm = normalizeForMatch(result.name)
-        if (!resultNorm) return false
-
-        const matched = tokens.filter((token) =>
-            resultNorm.includes(token),
-        ).length
-        const titleMatch = matched / tokens.length >= TITLE_MATCH_THRESHOLD
-        if (!titleMatch) return false
-
-        if (trackSec === null || !result.durationInSec) return true
-        return Math.abs(result.durationInSec - trackSec) <= 30
-    })
-}
-
-function normalizeForMatch(value: string): string {
-    return value
-        .toLowerCase()
-        .replace(/[^a-z0-9 ]/g, '')
-        .replace(/ {2,}/g, ' ')
-        .trim()
-}
-
-export function parseDurationString(duration?: string): number | null {
-    if (!duration) return null
-    const parts = duration.split(':').map(Number)
-    if (parts.some(Number.isNaN)) return null
-    if (parts.length === 2) return parts[0] * 60 + parts[1]
-    if (parts.length === 3) return parts[0] * 3600 + parts[1] * 60 + parts[2]
-    return null
-}
+export { streamViaYtDlp, streamViaYtDlpSearch, createResilientStream } from './streamBridge'
+export { streamViaSoundCloud, findMatchingSoundCloudResult, parseDurationString } from './soundcloudMatcher'
